example_repo/requirements-dev.txt
--extra-index-url https://esc-982534387049.d.codeartifact.us-east-1.amazonaws.com/pypi/esc-python/simple

annotated-types==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
asn1crypto==1.5.1 ; python_version >= "3.11" and python_version < "4.0"
aws-arns==1.5.1 ; python_version >= "3.11" and python_version < "4.0"
aws-config==0.1.4 ; python_version >= "3.11" and python_version < "4.0"
aws-console-url==1.4.1 ; python_version >= "3.11" and python_version < "4.0"
backports-tarfile==1.2.0 ; python_version == "3.11" and platform_machine != "ppc64le" and platform_machine != "s390x"
boto-session-manager==1.8.1 ; python_version >= "3.11" and python_version < "4.0"
boto3==1.40.21 ; python_version >= "3.11" and python_version < "4.0"
botocore==1.40.21 ; python_version >= "3.11" and python_version < "4.0"
build==1.3.0 ; python_version >= "3.11" and python_version < "4.0"
certifi==2025.8.3 ; python_version >= "3.11" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.11" and python_version < "4.0" and (platform_machine != "ppc64le" and platform_machine != "s390x" and sys_platform == "linux" and platform_python_implementation != "PyPy" or implementation_name == "pypy") and python_version >= "3.8"
charset-normalizer==3.4.3 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and (os_name == "nt" or sys_platform == "win32")
configcraft==0.1.2 ; python_version >= "3.11" and python_version < "4.0"
cookiecutter-pywf-internal-proprietary-demo==0.1.2 ; python_version >= "3.11" and python_version < "4.0"
cryptography==45.0.6 ; python_version >= "3.11" and python_version < "4.0" and platform_machine != "ppc64le" and platform_machine != "s390x" and sys_platform == "linux"
docutils==0.21.2 ; python_version >= "3.11" and python_version < "4.0"
enum-mate==0.1.1 ; python_version >= "3.11" and python_version < "4.0"
events==0.5 ; python_version >= "3.11" and python_version < "4.0"
func-args==1.0.1 ; python_version >= "3.11" and python_version < "4.0"
greenlet==3.2.4 ; python_version >= "3.11" and python_version < "3.14" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
id==1.5.0 ; python_version >= "3.11" and python_version < "4.0"
idna==3.10 ; python_version >= "3.11" and python_version < "4.0"
importlib-metadata==8.7.0 ; python_version == "3.11" and platform_machine != "ppc64le" and platform_machine != "s390x"
iterproxy==0.3.1 ; python_version >= "3.11" and python_version < "4.0"
jaraco-classes==3.4.0 ; python_version >= "3.11" and python_version < "4.0" and platform_machine != "ppc64le" and platform_machine != "s390x"
jaraco-context==6.0.1 ; python_version >= "3.11" and python_version < "4.0" and platform_machine != "ppc64le" and platform_machine != "s390x"
jaraco-functools==4.3.0 ; python_version >= "3.11" and python_version < "4.0" and platform_machine != "ppc64le" and platform_machine != "s390x"
jeepney==0.9.0 ; python_version >= "3.11" and python_version < "4.0" and platform_machine != "ppc64le" and platform_machine != "s390x" and sys_platform == "linux"
jmespath==1.0.1 ; python_version >= "3.11" and python_version < "4.0"
keyring==25.6.0 ; python_version >= "3.11" and python_version < "4.0" and platform_machine != "ppc64le" and platform_machine != "s390x"
markdown-it-py==4.0.0 ; python_version >= "3.11" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.11" and python_version < "4.0"
more-itertools==10.7.0 ; python_version >= "3.11" and python_version < "4.0" and platform_machine != "ppc64le" and platform_machine != "s390x"
nh3==0.3.0 ; python_version >= "3.11" and python_version < "4.0"
opensearch-py==2.8.0 ; python_version >= "3.11" and python_version < "4.0"
packaging==25.0 ; python_version >= "3.11" and python_version < "4.0"
pathlib-mate==1.3.2 ; python_version >= "3.11" and python_version < "4.0"
pg8000==1.31.4 ; python_version >= "3.11" and python_version < "4.0"
polars==1.32.3 ; python_version >= "3.11" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.11" and python_version < "4.0" and (platform_machine != "ppc64le" and platform_machine != "s390x" and sys_platform == "linux" and platform_python_implementation != "PyPy" or implementation_name == "pypy") and python_version >= "3.8"
pydantic-core==2.33.2 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.11.7 ; python_version >= "3.11" and python_version < "4.0"
pygments==2.19.2 ; python_version >= "3.11" and python_version < "4.0"
pynamodb-mate==6.0.1.1 ; python_version >= "3.11" and python_version < "4.0"
pynamodb==6.1.0 ; python_version >= "3.11" and python_version < "4.0"
pyproject-hooks==1.2.0 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_version < "4.0"
pywin32-ctypes==0.2.3 ; python_version >= "3.11" and python_version < "4.0" and platform_machine != "ppc64le" and platform_machine != "s390x" and sys_platform == "win32"
readme-renderer==44.0 ; python_version >= "3.11" and python_version < "4.0"
requests-toolbelt==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
requests==2.32.5 ; python_version >= "3.11" and python_version < "4.0"
rfc3986==2.0.0 ; python_version >= "3.11" and python_version < "4.0"
rich==13.9.4 ; python_version >= "3.11" and python_version < "4.0"
s3manifesto==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
s3pathlib==2.3.6 ; python_version >= "3.11" and python_version < "4.0"
s3transfer==0.13.1 ; python_version >= "3.11" and python_version < "4.0"
scramp==1.4.6 ; python_version >= "3.11" and python_version < "4.0"
secretstorage==3.3.3 ; python_version >= "3.11" and python_version < "4.0" and platform_machine != "ppc64le" and platform_machine != "s390x" and sys_platform == "linux"
simple-aws-ssm-parameter-store==0.2.3 ; python_version >= "3.11" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.11" and python_version < "4.0"
smart-open==7.3.0.post1 ; python_version >= "3.11" and python_version < "4.0"
sqlalchemy==2.0.43 ; python_version >= "3.11" and python_version < "4.0"
twine==6.1.0 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.15.0 ; python_version >= "3.11" and python_version < "4.0"
typing-inspection==0.4.1 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.5.0 ; python_version >= "3.11" and python_version < "4.0"
vislog==0.1.2 ; python_version >= "3.11" and python_version < "4.0"
wheel==0.45.1 ; python_version >= "3.11" and python_version < "4.0"
which-bsm==0.1.1 ; python_version >= "3.11" and python_version < "4.0"
which-env==0.1.1 ; python_version >= "3.11" and python_version < "4.0"
which-runtime==0.1.1 ; python_version >= "3.11" and python_version < "4.0"
wrapt==1.17.3 ; python_version >= "3.11" and python_version < "4.0"
zipp==3.23.0 ; python_version == "3.11" and platform_machine != "ppc64le" and platform_machine != "s390x"
